+++ packages/core-data/src/reducer.js
@@ -234,6 +234,45 @@ export function hasUploadPermissions( state = true, action ) {
 	return state;
 }
 
+/**
+ * Reducer returning the most recent autosave.
+ *
+ * @param  {Object} state  Current state.
+ * @param  {Object} action Dispatched action.
+ *
+ * @return {Object} Updated state.
+ */
+export function autosave( state = {}, action ) {
+	switch ( action.type ) {
+		case 'RESET_AUTOSAVE':
+			const { postId, autosave: autosavePost } = action;
+			const [ title, excerpt, content ] = [
+				'title',
+				'excerpt',
+				'content',
+			].map( ( field ) => {
+				// TODO - this is a duplication of editor/store/reducer #getPostRawValue. Unduplicate.
+				const value = autosavePost[ field ];
+				if ( value && 'object' === typeof value && 'raw' in value ) {
+					return value.raw;
+				}
+
+				return value;
+			} );
+
+			return {
+				[ postId ]: {
+					title,
+					excerpt,
+					content,
+				},
+				...state,
+			};
+	}
+
+	return state;
+}
+
 export default combineReducers( {
 	terms,
 	users,
@@ -242,4 +281,5 @@ export default combineReducers( {
 	entities,
 	embedPreviews,
 	hasUploadPermissions,
+	autosave,
 } );