+++ packages/editor/src/store/effects/posts.js
@@ -12,13 +12,15 @@ import { __ } from '@wordpress/i18n';
 // TODO: Ideally this would be the only dispatch in scope. This requires either
 // refactoring editor actions to yielded controls, or replacing direct dispatch
 // on the editor store with action creators (e.g. `REQUEST_POST_UPDATE_START`).
-import { dispatch as dataDispatch } from '@wordpress/data';
+import {
+	dispatch as dataDispatch,
+	select as dataSelect,
+} from '@wordpress/data';
 
 /**
  * Internal dependencies
  */
 import {
-	resetAutosave,
 	resetPost,
 	updatePost,
 } from '../actions';
@@ -26,7 +28,6 @@ import {
 	getCurrentPost,
 	getPostEdits,
 	getEditedPostContent,
-	getAutosave,
 	getCurrentPostType,
 	isEditedPostSaveable,
 	isEditedPostNew,
@@ -103,11 +104,13 @@ export const requestPostUpdate = async ( action, store ) => {
 
 	let request;
 	if ( isAutosave ) {
+		const { getAutosave } = dataSelect( 'core' );
+
 		// Ensure autosaves contain all expected fields, using autosave or
 		// post values as fallback if not otherwise included in edits.
 		toSend = {
 			...pick( post, [ 'title', 'content', 'excerpt' ] ),
-			...getAutosave( state ),
+			...getAutosave( post.id ),
 			...toSend,
 		};
 
@@ -129,8 +132,11 @@ export const requestPostUpdate = async ( action, store ) => {
 
 	try {
 		const newPost = await request;
-		const reset = isAutosave ? resetAutosave : resetPost;
-		dispatch( reset( newPost ) );
+		if ( isAutosave ) {
+			dataDispatch( 'core' ).resetAutosave( post.id, newPost );
+		} else {
+			dispatch( resetPost( newPost ) );
+		}
 
 		// An autosave may be processed by the server as a regular save
 		// when its update is requested by the author and the post was
@@ -317,3 +323,4 @@ export const refreshPost = async ( action, store ) => {
 	} );
 	dispatch( resetPost( newPost ) );
 };
+